pyproject.toml
[tool.poetry]
name = "insights"
version = "0.1.0"
description = ""
authors = ["Amer Alnajdawi <amernajdawi8@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.14"
streamlit = "^1.35.0"
streamlit-chat = "^0.1.1"
pdfplumber="0.9.0"
pandas="2.2.2"
openai="1.41.0"
pydantic = "^2.6.1"
langchain="0.2.15"
qdrant-client="1.11.1"
nltk="3.9.1"
tomotopy="0.13.0"
pyvis="0.3.2"
langchain-openai="^0.1.1"  
python-dotenv="^1.0.0"  
selenium="4.24.0"
langchain-community="0.2.14"
langchain-qdrant = "^0.1.3"
beautifulsoup4 = "^4.12.3"






[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
